Try previous patch versions when searching for genX.Y.Z directory.
In the future we may not generate a low-level wrapper for each patch
version. In particular not when it would be identical to the previous
one. With this change COPT.jl will use the genX.Y.Z directory of a
previous patch version if there is no matching one for the loaded
library.